# lib/api-zod/src/generated/types/awardXpBody.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */

export interface AwardXpBody {
  xp: number;
  subject: string;
  reason?: string;
}