# lib/api-zod/src/generated/types/progressSummary.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */
import type { SubjectProgress } from "./subjectProgress";

export interface ProgressSummary {
  deviceId: string;
  totalXp: number;
  level: number;
  streak: number;
  totalQuestions: number;
  subjectBreakdown: SubjectProgress[];
  badges: string[];
  lastActiveAt?: Date;
}