# lib/api-zod/src/generated/types/studentProfile.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */

export interface StudentProfile {
  id: number;
  deviceId: string;
  name: string;
  grade: string;
  board: string;
  subjects: string[];
  goal?: string;
  preferredLanguage?: string;
  voicePersonality?: string;
  createdAt: Date;
  updatedAt: Date;
}