# lib/api-zod/src/generated/types/subjectProgress.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */

export interface SubjectProgress {
  subject: string;
  questionCount: number;
  xp: number;
}