# lib/api-zod/src/generated/types/upsertProfileBody.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */

export interface UpsertProfileBody {
  deviceId: string;
  name: string;
  grade: string;
  board: string;
  subjects: string[];
  goal?: string;
  preferredLanguage?: string;
  voicePersonality?: string;
}